realtime-gateway/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
click==8.2.1
ecdsa==0.19.1
fastapi==0.115.13
greenlet==3.2.3
h11==0.16.0
idna==3.10
jose==1.0.0
kafka-python==2.2.13
pyasn1==0.6.1
pydantic==2.11.7
pydantic_core==2.33.2
PyMySQL==1.1.1
python-jose==3.5.0
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.41
starlette==0.46.2
typing-inspection==0.4.1
typing_extensions==4.14.0
uvicorn==0.34.3
websockets==15.0.1
